wip: move last mutations to db layer
