+++ __tests-pacts__/serlo.org-database-layer/threads/thread-set-archive.ts
@@ -21,13 +21,13 @@
  */
 import { gql } from 'apollo-server'
 
-import { article, comment3, user } from '../../__fixtures__'
-import { createTestClient } from '../../__tests__/__utils__'
-import { mockEndpointsForThreads } from '../../__tests__/schema/thread/thread'
+import { article, comment3, user } from '../../../__fixtures__'
+import { createTestClient } from '../../../__tests__/__utils__'
+import { mockEndpointsForThreads } from '../../../__tests__/schema/thread/thread'
 import {
   addMutationInteraction,
   assertSuccessfulGraphQLMutation,
-} from '../__utils__'
+} from '../../__utils__'
 import { encodeThreadId } from '~/schema/thread'
 
 test('/thread/set-archive', async () => {
