+++ package.json
@@ -25,10 +25,9 @@
     "lint:eslint": "yarn _eslint",
     "lint:prettier": "yarn _prettier --check",
     "lint:tsc": "tsc --noEmit",
-    "pacts": "npm-run-all pacts:clean pacts:serlo-org pacts:serlo-org-database-layer",
+    "pacts": "npm-run-all pacts:clean pacts:serlo-org-database-layer",
     "pacts:clean": "rimraf pacts",
     "pacts:publish": "ts-node scripts/publish-pacts",
-    "pacts:serlo-org": "jest --config jest.config-pacts-serlo-org.js --runInBand",
     "pacts:serlo-org-database-layer": "jest --config jest.config-pacts-serlo-org-database-layer.js --runInBand",
     "preregenerate-types": "graphql-codegen --config codegen.yml",
     "regenerate-types": "ts-node scripts/regenerate-types-package",
